backend/requirements.txt
flask
pillow
werkzeug
numpy
